Full redesign

+++ app/controllers/likes_controller.rb
@@ -1,9 +1,11 @@
 class LikesController < ApplicationController
+	after_action :verify_authorized
 
 	def create
 		@like = Like.new
 		@like.entity = context
 		@like.author = current_user
+		authorize @like
 		unless @like.entity.liked? current_user
 			@like.save
 			redirect_to :back
@@ -14,10 +16,11 @@ class LikesController < ApplicationController
 	end
 
 	def destroy
+		authorize Like.new
 		if context.liked? current_user
 			like = Like.where(entity: context, author: current_user).take
 			like.destroy
-			redirect_to :back
+			redirect_to :back,  status: 303  
 		else
 			head :bad_request
 		end
